requirements.txt
streamlit
pandas
matplotlib
plotly
altair
python-dotenv
openai